+++ package.json
@@ -11,7 +11,7 @@
     "jsrsasign": "5.0.1",
     "keymirror": "~0.1.0",
     "react": "^15.1.0",
-    "react-native": "0.27.2",
+    "react-native": "0.31",
     "react-native-button": "*",
     "react-native-vector-icons": "^2.0.3",
     "react-redux": "^4.4.5",
